Make a distinction between immutable/mutable data.
"Mutable" is, in this context, a synonym for "type-checked". We don't
care that much if already-stored data is corrupt (dealing with that is
left as an exercise for the application), but new data should always
adhere to the schema. Thus, to make modifications to either a List or a
Table, one must use .As to acquire a reference to a
MutableList/MutableTable.

This has the added advantage that, in the future, changes to a
MutableList/MutableTable can be tracked within the mutable bit and
applied as a diff to the backing store in cases where that'd make a
difference. But that is neither today nor tonight since it probably
won't make a difference for a long while.